# Speed up user-credentials output