--- a/tools/qontract_cli.py
+++ b/tools/qontract_cli.py
@@ -918,42 +918,42 @@ def clusters_aws_account_ids(ctx):
 @get.command()
 @click.pass_context
 def terraform_users_credentials(ctx) -> None:
-    accounts = queries.get_state_aws_accounts()
+    credentials = []
     state = init_state(integration="account-notifier")
 
     skip_accounts, appsre_pgp_key, _ = tfu.get_reencrypt_settings()
 
-    accounts, working_dirs, _, aws_api = tfu.setup(
-        False,
-        1,
-        skip_accounts,
-        account_name=None,
-        appsre_pgp_key=appsre_pgp_key,
-    )
+    if skip_accounts:
+        accounts, working_dirs, _, aws_api = tfu.setup(
+            False,
+            1,
+            skip_accounts,
+            account_name=None,
+            appsre_pgp_key=appsre_pgp_key,
+        )
 
-    tf = Terraform(
-        tfu.QONTRACT_INTEGRATION,
-        tfu.QONTRACT_INTEGRATION_VERSION,
-        tfu.QONTRACT_TF_PREFIX,
-        accounts,
-        working_dirs,
-        10,
-        aws_api,
-        init_users=True,
-    )
-    credentials = []
-    for account, output in tf.outputs.items():
-        if account in skip_accounts:
-            user_passwords = tf.format_output(output, tf.OUTPUT_TYPE_PASSWORDS)
-            console_urls = tf.format_output(output, tf.OUTPUT_TYPE_CONSOLEURLS)
-            for user_name, enc_password in user_passwords.items():
-                item = {
-                    "account": account,
-                    "console_url": console_urls[account],
-                    "user_name": user_name,
-                    "encrypted_password": enc_password,
-                }
-                credentials.append(item)
+        tf = Terraform(
+            tfu.QONTRACT_INTEGRATION,
+            tfu.QONTRACT_INTEGRATION_VERSION,
+            tfu.QONTRACT_TF_PREFIX,
+            accounts,
+            working_dirs,
+            10,
+            aws_api,
+            init_users=True,
+        )
+        for account, output in tf.outputs.items():
+            if account in skip_accounts:
+                user_passwords = tf.format_output(output, tf.OUTPUT_TYPE_PASSWORDS)
+                console_urls = tf.format_output(output, tf.OUTPUT_TYPE_CONSOLEURLS)
+                for user_name, enc_password in user_passwords.items():
+                    item = {
+                        "account": account,
+                        "console_url": console_urls[account],
+                        "user_name": user_name,
+                        "encrypted_password": enc_password,
+                    }
+                    credentials.append(item)
 
     secrets = state.ls()